Increase prime_multipliers list
This will need to be refactored to refer to number_list. As is stands now, it may not work for larger numbers. The reason for the increase was that for 168/1000 primes, the wrong last number was returned, meaning there were still some composite numbers in the list which need to be removed.